[DR-2] fix: eslint issue
