+++ fe/Douren-frontend/.eslintrc.cjs
@@ -20,7 +20,6 @@ module.exports = {
       "warn",
       { allowConstantExport: true },
     ],
-    "@typescript-eslint/no-unused-vars": "off",
     "react-hooks/exhaustive-deps": 0,
     "no-empty-pattern": "off",
     "@typescript-eslint/no-explicit-any": "off",
