+++ fe/Douren-frontend/src/routes/event/-subcomponent/FilterContainer.tsx
@@ -1,11 +1,9 @@
 import classNames from "classnames/bind";
 import styles from "../EventPage.module.css";
-import SortSelect from "../../../components/SortSelect/SortSelect";
 import { TagFilter } from "@/components/TagFilter/TagFilter.tsx";
 import SelectComponent from "../../../components/SelectComponent/SelectComponent";
-import { useState } from "react";
 import { useSortSelectContext } from "@/context/SortSelectContext/useSortSelectContext.ts";
-import { useUpdatePageSideEffect } from "@/context/DataOperationContext";
+import { useUpdatePageSideEffect } from "@/context/DataOperationContext/useUpdateSideEffect.ts";
 
 const FilterContainer = () => {
   const sx = classNames.bind(styles);
